requirements.txt
streamlit==1.24.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
pandas==2.0.3